Move top level state machine into state machines dir
